parsers: add regex parser
This is meant to be the basic parser using regexes. It replaces the old
inline code parsing fields based on their names. It allows explicitly
specifying how to parse given field.

It basically works the same way as the old parsing methods. One minor
change is preserving order of matches if more than 1 unique value was
found. For fields defined using the old syntax order is not maintained
to provide full backward compatibility.

Signed-off-by: Rafał Miłecki <rafal@milecki.pl>