Added area support for fields
Fields specifying a parser now also accept an "area" argument. This allows one to specify the page(s) and coordinates of the area to be extracted and then applies the specified parser to the results as normal :)
Currently this only works for pdftotext.
Incorporated changes into all existing functions where possible to allow for ease of future maintainability.
Tutorial was updated to show new option and its syntax.