Modified as previously not working
I was unable to get the previous to_csv.py file to work. It was trying to find keys in all values, even if they were not dictionaries.

I modified the code to now check if the field is a list of dictionaries first. Only then does it start looking for keys and values.
All values where the name starts or ends with "date" is still attempted to be parsed as a date format.