Adjusting lines plugin to only look for line between first and last
Currently the behaviour of the lines plugin seems a little counter intuitive in that it will record "line" even before a "first_line" has been found.

With these changes, the plugin will begin by finding a "first_line". When that matches, it will look for "line", appending matches to the first_line until it either finds "last_line" or another "first_line".

This allows for more logical structuring of text searches and control of where and what data is extracted.