Be more verbose about loaded templates (#355)
People seem to be often confused by invoice2data CLI behaviour. When
something doesn't work the default output doesn't seem to be enough to
understand what's happening.

Try to improve that situation be printing info about templates that were
loaded and the used one. Use info level messages for that basic stuff so
it's clear what is happening even without the --debug switch.

Signed-off-by: Rafał Miłecki <rafal@milecki.pl>

Co-authored-by: Rafał Miłecki <rafal@milecki.pl>
Co-authored-by: Manu <3916435+m3nu@users.noreply.github.com>