Add posibility to parse multiple line definitions bosd
refactored code to allow multiple lines to be parsed.

Restructuring was needed.

As we want to output the lines in the order of the invoice file.

Independent of the order of the template file.

Example yaml input

lines:

 - start: Barcode

   end:  Netto totaal

   line: (?P<line_note>(FOOD))

 - start: Barcode

   end:  Netto totaal

   line: (?P<description>(.+))\s+(?P<price>\d+\d+)

it is also possible to contentate multi line tags.

alternative example sammymaystone.

lines:

 - start: Item\s+Quantity\s+Rate\s+Amount

   end:  Subtotal

   first_line: 'Service (?P<item>\w)\s+(?P<qty>\S+)\s+.(?P<unitprice>\d+.\d{2})\s+.?(?P<linetotal>\d+.\d{2})'

   line: '(?P<item>.*)'

   skip_line: ['Pino \w'] # 'Description:', Notes:,

   last_line: '(?P<desc>Parts:.*)'

   types:

     qty: int

     unitprice: float

     linetotal: float

input:

Service A                                                                 12                      0.00         20.00

Description: Repair

Notes: Replaced capacitor

Parts: 1 x cap_a

Tax: 0.2%

output:

{'item': 'A\nDescription: Repair\nNotes: Replaced capacitor', 'qty': 12, 'unitprice': 10.0, 'linetotal': 120.0, 'desc': 'Parts: 1 x cap_a'}